backend/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
flask==3.1.0
flask-cors==5.0.1
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
werkzeug==3.1.3
zipp==3.21.0
